added the function to the views.py
